Add to best fit is fucked. Invstgating

+++ SDK_C++/CodeCraft-2021/vm_data_manager.cpp
@@ -29,7 +29,7 @@ VmInfo& VmDataManager::GetVm(const std::string& vm_name) {
     return vm_info_list_[index_vm_name_[vm_name]];
 }
 
-int VmDataManager::GetVmId(const std::string& vm_name) {
+uint16_t VmDataManager::GetVmId(const std::string& vm_name) {
     return index_vm_name_[vm_name];
 }
